made storage and methods for user data
